Tratamento de erros

+++ src/Pages/Repositories/index.js
@@ -1,15 +1,22 @@
 import React, { useEffect, useState }  from 'react';
 import * as S from './styled';
+import { useHistory } from 'react-router-dom';
 
 export default function Repositories(){
     const [repositories, setRepositories] = useState([]);
+    const history = useHistory();
 
     useEffect(() => {
         let repositoriesName = localStorage.getItem('repositoriesName');
-        repositoriesName = JSON.parse(repositoriesName);
-        console.log(repositoriesName);
-        setRepositories(repositoriesName);
-        localStorage.clear();
+        if(repositoriesName != null){
+            repositoriesName = JSON.parse(repositoriesName);
+            setRepositories(repositoriesName);
+            localStorage.clear();
+        }
+        else {
+            history.pushState('/');
+        }
+        
     }, []);
     
     return (
